MFAWebApp/requirements.txt
deepface==0.0.79
easydict==1.10
facenet_pytorch==2.5.3
Flask==2.2.3
flask_cors==3.0.10
#mysql_connector_repackaged==0.3.1
mysql-connector-python==8.0.28
numpy==1.24.2
Pillow==9.5.0
scikit-image==0.18.3
tensorboardX==2.6
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
